pyproject.toml
[project]
name = "ai-projects2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mcp>=1.26.0",
    "pydantic-ai>=1.52.0",
    "pytest>=9.0.2",
    "weaviate-client>=4.19.2",
]
